Make deleteButton and Backspace key work properly
Clicking deleteButton or keydown Backspace would cause more numbers to
appear instead of backspacing the last number.

This was because display.textContent was showing
roundResult(firstNumber), but firstNumber was still in the longer form.
backspace() backspaced the last number of firstNumber, not
roundResult(firstNumber), and displayed the backspaced firstNumber,
causing it to seem like the displayed number became longer.